Fixed data type for balance in account class

+++ src/main/java/bank/Account.java
@@ -3,9 +3,9 @@ package bank;
 public class Account {
   private int id;
   private String type;
-  private float balance;
+  private double balance;
 
-  public Account(int id, String type, float balance) {
+  public Account(int id, String type, double balance) {
     setId(id);
     setType(type);
     setBalance(balance);
@@ -27,11 +27,11 @@ public class Account {
     this.type = type;
   }
 
-  public float getBalance() {
+  public double getBalance() {
     return this.balance;
   }
 
-  public void setBalance(float balance) {
+  public void setBalance(double balance) {
     this.balance = balance;
   }
